requirements.txt
Django==2.0.1
django-debug-toolbar==1.9.1
django-rest-framework==0.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
mock==2.0.0
pbr==3.1.1
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
sqlparse==0.2.4
